% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_created}
\alias{add_created}
\title{Add file created date}
\usage{
add_created(.data)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with file created dates
}
\description{
Adds or overwrites a column named "created", containing the file created date
of each file.
}
\seealso{
\code{\link{get_created}}
}
