% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_directoryname}
\alias{add_directoryname}
\title{Add directory name}
\usage{
add_directoryname(.data)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with directory names
}
\description{
Adds or overwrites a column named "directoryname", containing the directory
of each file.
}
\seealso{
\code{\link{get_directoryname}}
}
