% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_dos}
\alias{add_dos}
\title{Add file hash for dos line-endings}
\usage{
add_dos(.data, ...)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with dos file hashes
}
\description{
Adds or overwrites a column named "dos", containing the file hash
for each file, as computed by the \code{get_dos} function.
}
\seealso{
\code{\link{get_dos}}
}
