% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_ext}
\alias{add_ext}
\title{Add file extension}
\usage{
add_ext(.data)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with file extensions
}
\description{
Adds or overwrites a column named "ext", containing the file extension
of each file.
}
\seealso{
\code{\link{get_ext}}
}
