% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_filename}
\alias{add_filename}
\title{Add filename}
\usage{
add_filename(.data)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with filenames
}
\description{
Adds or overwrites a column named "filename", containing the filename (without
directory) of the path for each file.
}
\seealso{
\code{\link{get_filename}}
}
