% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_files}
\alias{add_files}
\title{Add files to fduper object path listing}
\usage{
add_files(.data, files)
}
\arguments{
\item{.data}{A fduper object}

\item{files}{A vector of file paths}
}
\value{
The modified fduper object with the additional paths added
}
\description{
Provide a character vector of file paths to add to the paths in the fduper object.
}
\examples{
f \%>\% add_files("~/my_file.txt")
}
