% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_hash}
\alias{add_hash}
\title{Add file hash}
\usage{
add_hash(.data, ...)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with file hashes
}
\description{
Adds or overwrites a column named "hash", containing the file hash
for each file, as computed by the \code{get_hash} function.
}
\seealso{
\code{\link{get_hash}}
}
