% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_modified}
\alias{add_modified}
\title{Add file modified date}
\usage{
add_modified(.data)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with file modified dates
}
\description{
Adds or overwrites a column named "modified", containing the file modification date
of each file.
}
\seealso{
\code{\link{get_modified}}
}
