% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_size}
\alias{add_size}
\title{Add file size in bytes}
\usage{
add_size(.data)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with file sizes in bytes
}
\description{
Adds or overwrites a column named "size", containing the file size in bytes
for each file.
}
\seealso{
\code{\link{get_size}}
}
