% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{add_unix}
\alias{add_unix}
\title{Add file hash for unix line-endings}
\usage{
add_unix(.data, ...)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The modified fduper object with unix file hashes
}
\description{
Adds or overwrites a column named "unix", containing the file hash
for each file, as computed by the \code{get_unix} function.
}
\seealso{
\code{\link{get_unix}}
}
