% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dplyr.R
\name{arrange.fduper}
\alias{arrange.fduper}
\title{fduper wrapper for dplyr arrange}
\usage{
\method{arrange}{fduper}(.data, ...)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The fduper object after processing
}
\description{
fduper wrapper for dplyr arrange
}
