% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fduper.R
\name{fduper}
\alias{fduper}
\title{fduper constructor}
\usage{
fduper(path = character(0), ...)
}
\arguments{
\item{path}{A character vector of paths to be added recursively}

\item{...}{Additional columns to add to the underlying fduper tibble}
}
\value{
A fduper object, to be piped to other methods for processing.
}
\description{
Create a fduper object to be passed to subsequent methods from the fduper
package.
}
\examples{
fduper()
}
