% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dplyr.R
\name{filter.fduper}
\alias{filter.fduper}
\title{fduper wrapper for dplyr filter}
\usage{
\method{filter}{fduper}(.data, ...)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The fduper object after processing
}
\description{
fduper wrapper for dplyr filter
}
