% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_created}
\alias{get_created}
\title{Get file created date}
\usage{
get_created(path)
}
\arguments{
\item{path}{The path to a file}
}
\value{
The file created date of the file
}
\description{
Returns the file created date of a file
}
