% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_directoryname}
\alias{get_directoryname}
\title{Get directory name}
\usage{
get_directoryname(path)
}
\arguments{
\item{path}{The path to a file}
}
\value{
The directory path, without the filename
}
\description{
Returns the directory name of a file
}
