% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dirs}
\alias{get_dirs}
\title{get_dirs: Collect directory paths}
\usage{
get_dirs(path, pattern = NULL, all.dirs = TRUE, recursive = TRUE,
  ignore.case = FALSE)
}
\description{
get_dirs: Collect directory paths
}
\examples{
get_dirs(sample_path(), recursive=TRUE)
get_dirs(sample_path(), recursive=FALSE)
}
