% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_dos_}
\alias{get_dos_}
\title{Get file hash of dos line-ending version of file}
\usage{
get_dos_(path, algo = "md5", default = get_hash(path))
}
\description{
Get file hash of dos line-ending version of file
}
