% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_ext}
\alias{get_ext}
\title{Get file extension}
\usage{
get_ext(path)
}
\arguments{
\item{path}{The path to a file}
}
\value{
The file extension as a string
}
\description{
Returns the file extension of a file
}
