% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_filename}
\alias{get_filename}
\title{Get filename}
\usage{
get_filename(path)
}
\arguments{
\item{path}{The path to a file}
}
\value{
The filename without the directory
}
\description{
Returns the basename of a file (the file name without the directory)
}
