% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_hash}
\alias{get_hash}
\title{Get file hash}
\usage{
get_hash(path, algo = "md5")
}
\arguments{
\item{path}{The path to a file}

\item{algo}{The hashing algorithm as a string. Default is "md5". See the
digest package for allowable values.}
}
\value{
The file hash as a string
}
\description{
Returns the hash of the file. Different hashing algoritms can be used,
with the default being md5sum. Underlying work is performed by the
digest package.
}
