% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_modified}
\alias{get_modified}
\title{Get file modified date}
\usage{
get_modified(path)
}
\arguments{
\item{path}{The path to a file}
}
\value{
The file modified date of the file
}
\description{
Returns the file modified date of a file
}
