% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_properties.R
\name{get_size}
\alias{get_size}
\title{Get file size in bytes}
\usage{
get_size(path)
}
\arguments{
\item{path}{The path to a file}
}
\value{
The file size of the file in bytes
}
\description{
Returns the file size date of a file in bytes
}
