% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-reducers.R
\name{identify}
\alias{identify}
\title{Label groups of items sharing certain values}
\usage{
identify(.data, ...)
}
\arguments{
\item{.data}{A fduper object}

\item{...}{Column names for grouping files as duplicates}
}
\value{
The modified fduper object with duplicate groups labelled
}
\description{
Groups the files in the fduper dataset by the columns pass in the dot arguments,
then adds a column "group" with a unique value for each set of duplicates.
}
\examples{
f \%>\% identify(size)
f \%>\% identify(size, hash)
}
