% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fduper.R
\name{is_fduper}
\alias{is_fduper}
\title{Determine if object is class fduper}
\usage{
is_fduper(.data)
}
\arguments{
\item{.data}{A object to be evaluated}
}
\value{
TRUE if object is of class fduper
}
\description{
Check if object contains the class "fduper".
}
