% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dplyr.R
\name{mutate.fduper}
\alias{mutate.fduper}
\title{fduper wrapper for dplyr mutate}
\usage{
\method{mutate}{fduper}(.data, ...)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The fduper object after processing
}
\description{
fduper wrapper for dplyr mutate
}
