% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fduper.R
\name{sample_path}
\alias{sample_path}
\title{Sample documents path for testing}
\usage{
sample_path()
}
\value{
A path to some test files within the package directory
}
\description{
For testing workflows, this sample document path can be used
}
