% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dplyr.R
\name{ungroup.fduper}
\alias{ungroup.fduper}
\title{fduper wrapper for dplyr ungroup}
\usage{
\method{ungroup}{fduper}(.data, ...)
}
\arguments{
\item{.data}{A fduper object}
}
\value{
The fduper object after processing
}
\description{
fduper wrapper for dplyr ungroup
}
