% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_dirs}
\alias{add_dirs}
\title{add_dirs: Add files in directories to fduper object}
\usage{
add_dirs(.data, path, recursive = TRUE, ...)
}
\description{
add_dirs: Add files in directories to fduper object
}
