% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fduper.R
\name{fduper}
\alias{fduper}
\alias{fduper}
\title{fduper: Eliminate Duplicate Files}
\usage{
fduper(path = character(0), ...)

fduper(path = character(0), ...)
}
\arguments{
\item{path}{A character vector of paths}

\item{...}{Additional columns to add to the tibble}
}
\value{
A fduper object
}
\description{
fduper: Eliminate Duplicate Files

fduper: Constructor
}
