% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dirs}
\alias{get_dirs}
\title{get_dirs: Collect directory paths}
\usage{
get_dirs(path, pattern = NULL, all.files = TRUE, recursive = TRUE, ...)
}
\description{
get_dirs: Collect directory paths
}
\examples{
get_dirs("~/data", recursive=TRUE)
get_dirs("~/data", recursive=FALSE)
get_dirs("~/source", "\\\\.git$", recursive=TRUE)
get_files(get_dirs("~/source", "\\\\.git$")) # for example, to remove from file list
}
