% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dirs}
\alias{get_dirs}
\title{Collect directory paths}
\usage{
get_dirs(path, pattern = NULL, all.dirs = TRUE, recursive = TRUE,
  ignore.case = FALSE)
}
\arguments{
\item{path}{A character vector of path names to search}

\item{pattern}{An optional regex pattern to match results against}

\item{all.dirs}{Whether or not to include directories starting with period, with default as TRUE.}

\item{recursive}{Whether or not to search path recursively, with default as TRUE}

\item{ignore.case}{Whether or not to use case-sensitive pattern matching, with default as FALSE}
}
\value{
A character vector containing the resulting directory paths
}
\description{
The \code{get_dirs} function addresses some of the quirks with list.files
and uses defaults that are more useful for fduper, such as \code{recursive = TRUE}
}
\examples{
get_dirs(sample_path(), recursive=TRUE)
get_dirs(sample_path(), recursive=FALSE)
}
