% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_files}
\alias{get_files}
\title{get_files: Collect file paths}
\usage{
get_files(path, pattern = NULL, all.files = TRUE, recursive = TRUE,
  ignore.case = FALSE)
}
\description{
get_files: Collect file paths
}
\examples{
get_files("~/data", recursive=TRUE)
get_files("~/data", recursive=FALSE)
get_files("~/data", "*.bz2$", recursive=TRUE)
}
