% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{keep_files_by_pattern}
\alias{keep_files_by_pattern}
\title{keep_files_by_pattern: Keep files by matching regex}
\usage{
keep_files_by_pattern(.data, files)
}
\description{
keep_files_by_pattern: Keep files by matching regex
}
