% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_dirs}
\alias{remove_dirs}
\title{remove_dir: Remove files in directories from fduper object}
\usage{
remove_dirs(.data, path, ...)
}
\description{
remove_dir: Remove files in directories from fduper object
}
