% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_files}
\alias{remove_files}
\title{remove_files: Remove character vector of files from fduper object}
\usage{
remove_files(.data, files)
}
\description{
remove_files: Remove character vector of files from fduper object
}
