% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_files_by_pattern}
\alias{remove_files_by_pattern}
\title{remove_files_by_pattern: Remove files by matching regex}
\usage{
remove_files_by_pattern(.data, files)
}
\description{
remove_files_by_pattern: Remove files by matching regex
}
